+++ unirest/src/main/java/kong/unirest/HttpRequestUniBody.java
@@ -25,6 +25,8 @@
 
 package kong.unirest;
 
+import kong.unirest.json.JSONElement;
+
 import java.nio.charset.Charset;
 import java.util.Optional;
 
@@ -44,8 +46,25 @@ class HttpRequestUniBody extends BaseRequest<RequestBodyEntity> implements Reque
 	}
 
 	@Override
+	public RequestBodyEntity body(JSONElement jsonBody) {
+		return body(jsonBody.toString());
+	}
+
+	@Override
+	public RequestBodyEntity body(Object objectBody) {
+		if(objectBody instanceof String){
+			return body((String)objectBody);
+		} else if (objectBody instanceof JsonNode){
+			return body((JsonNode) objectBody);
+		} else if (objectBody instanceof JSONElement){
+			return body(objectBody.toString());
+		}
+		return body(getObjectMapper().writeValue(objectBody));
+	}
+
+	@Override
 	public RequestBodyEntity body(String bodyAsString) {
-		this.body = new UnibodyString(bodyAsString, charSet);
+		this.body = new UnibodyString(bodyAsString);
 		return this;
 	}