+++ unirest/src/test/java/BehaviorTests/AsFileTest.java
@@ -25,6 +25,7 @@
 
 package BehaviorTests;
 
+import kong.unirest.HttpResponse;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
 import kong.unirest.JacksonObjectMapper;
@@ -32,27 +33,29 @@ import kong.unirest.TestUtil;
 import kong.unirest.Unirest;
 
 import java.io.File;
+import java.nio.file.FileAlreadyExistsException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.nio.file.StandardCopyOption;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public class AsFileTest extends BddTest {
+class AsFileTest extends BddTest {
 
-    private Path test = Paths.get("results.json");
-    private JacksonObjectMapper om = new JacksonObjectMapper();
+    private final Path test = Paths.get("results.json");
+    private final JacksonObjectMapper om = new JacksonObjectMapper();
 
     @Override @AfterEach
     public void tearDown() {
         try {
             Files.delete(test);
-        } catch (Exception e) { }
+        } catch (Exception ignored) { }
     }
 
     @Test
-    public void canSaveContentsIntoFile() {
+    void canSaveContentsIntoFile() {
         File result = Unirest.get(MockServer.GET)
                 .queryString("talking","heads")
                 .queryString("param3", "こんにちは")
@@ -68,7 +71,7 @@ public class AsFileTest extends BddTest {
     }
 
     @Test
-    public void canSaveContentsIntoFileAsync() throws Exception {
+    void canSaveContentsIntoFileAsync() throws Exception {
         File result = Unirest.get(MockServer.GET)
                 .queryString("talking","heads")
                 .queryString("param3", "こんにちは")
@@ -85,7 +88,7 @@ public class AsFileTest extends BddTest {
     }
 
     @Test
-    public void canSaveContentsIntoFileAsyncWithCallback() throws Exception {
+    void canSaveContentsIntoFileAsyncWithCallback() throws Exception {
         Unirest.get(MockServer.GET)
                 .queryString("talking","heads")
                 .queryString("param3", "こんにちは")
@@ -102,7 +105,7 @@ public class AsFileTest extends BddTest {
     }
 
     @Test
-    public void canDownloadABinaryFile() throws Exception {
+    void canDownloadABinaryFile() throws Exception {
         File f1 = TestUtil.rezFile("/spidey.jpg");
 
         File f2 = Unirest.get(MockServer.BINARYFILE)
@@ -110,5 +113,45 @@ public class AsFileTest extends BddTest {
                 .getBody();
 
         assertTrue(com.google.common.io.Files.equal(f1, f2));
+    }
+
+    @Test
+    void byDefaultFailWhenAttemptingToOverride() {
+        Unirest.get(MockServer.BINARYFILE)
+                .asFile(test.toString());
+
+        HttpResponse<File> f2 = Unirest.get(MockServer.BINARYFILE)
+                .asFile(test.toString());
+
+        assertTrue(f2.getParsingError().isPresent());
+        assertTrue(f2.getParsingError().get().getCause().getCause() instanceof FileAlreadyExistsException);
+    }
+
+    @Test
+    void canOverrideExistingFiles() {
+        File f1 = Unirest.get(MockServer.BINARYFILE)
+                .asFile(test.toString())
+                .getBody();
+
+        File f2 = Unirest.get(MockServer.BINARYFILE)
+                .asFile(test.toString(), StandardCopyOption.REPLACE_EXISTING)
+                .getBody();
+
+        assertEquals(f1, f2);
+    }
+
+    @Test
+    void canOverrideExistingFiles_Async() throws Exception {
+        File f1 = Unirest.get(MockServer.BINARYFILE)
+                .asFileAsync(test.toString())
+                .get()
+                .getBody();
+
+        File f2 = Unirest.get(MockServer.BINARYFILE)
+                .asFileAsync(test.toString(), StandardCopyOption.REPLACE_EXISTING)
+                .get()
+                .getBody();
+
+        assertEquals(f1, f2);
     }
 }