+++ unirest/src/test/java/BehaviorTests/AsObjectTest.java
@@ -30,6 +30,7 @@ import com.google.gson.Gson;
 import kong.unirest.*;
 import org.junit.jupiter.api.Test;
 
+import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
 import java.util.List;
 
@@ -37,16 +38,16 @@ import static org.hamcrest.CoreMatchers.instanceOf;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
 
-public class AsObjectTest extends BddTest {
+class AsObjectTest extends BddTest {
 
     @Test
-    public void basicJsonObjectMapperIsTheDefault() {
+    void basicJsonObjectMapperIsTheDefault() {
         Unirest.config().shutDown(true);
         assertThat(Unirest.config().getObjectMapper(), instanceOf(JsonObjectMapper.class));
     }
 
     @Test
-    public void basicJsonObjectMapperIsGoodEnough() {
+    void basicJsonObjectMapperIsGoodEnough() {
         Unirest.config().shutDown(true);
         MockServer.setJsonAsResponse(new Foo("bar"));
 
@@ -58,7 +59,7 @@ public class AsObjectTest extends BddTest {
     }
 
     @Test
-    public void whenNoBodyIsReturned() {
+    void whenNoBodyIsReturned() {
         HttpResponse<RequestCapture> i = Unirest.get(MockServer.NOBODY).asObject(RequestCapture.class);
 
         assertEquals(200, i.getStatus());
@@ -66,7 +67,7 @@ public class AsObjectTest extends BddTest {
     }
 
     @Test
-    public void canGetObjectResponse() {
+    void canGetObjectResponse() {
          Unirest.get(MockServer.GET)
                 .queryString("foo", "bar")
                 .asObject(RequestCapture.class)
@@ -75,7 +76,7 @@ public class AsObjectTest extends BddTest {
     }
 
     @Test
-    public void canGetObjectResponseAsync() throws Exception {
+    void canGetObjectResponseAsync() throws Exception {
         Unirest.get(MockServer.GET)
                 .queryString("foo", "bar")
                 .asObjectAsync(RequestCapture.class)
@@ -85,7 +86,7 @@ public class AsObjectTest extends BddTest {
     }
 
     @Test
-    public void canGetObjectResponseAsyncWithCallback() {
+    void canGetObjectResponseAsyncWithCallback() {
         Unirest.get(MockServer.GET)
                 .queryString("foo", "bar")
                 .asObjectAsync(RequestCapture.class, r -> {
@@ -98,7 +99,7 @@ public class AsObjectTest extends BddTest {
     }
 
     @Test
-    public void canPassAnObjectMapperAsPartOfARequest(){
+    void canPassAnObjectMapperAsPartOfARequest(){
         Unirest.config().setObjectMapper(null);
 
         TestingMapper mapper = new TestingMapper();
@@ -121,7 +122,37 @@ public class AsObjectTest extends BddTest {
     }
 
     @Test
-    public void ifTheObjectMapperFailsReturnEmptyAndAddToParsingError() {
+    void canOverrideBody() {
+        Unirest.post(MockServer.POST)
+                .body(new Foo("Apple"))
+                .body(new Foo("Orange"))
+                .asObject(RequestCapture.class)
+                .getBody()
+                .assertBody("{\"bar\":\"Orange\"}");
+    }
+
+    @Test
+    void setCharSetAfterBody() {
+        Unirest.post(MockServer.POST)
+                .body(new Foo("Orange"))
+                .charset(StandardCharsets.US_ASCII)
+                .asObject(RequestCapture.class)
+                .getBody()
+                .assertContentType("text/plain; charset=US-ASCII");
+    }
+
+    @Test
+    void setNoCharSetAfterBody() {
+        Unirest.post(MockServer.POST)
+                .body(new Foo("Orange"))
+                .noCharset()
+                .asObject(RequestCapture.class)
+                .getBody()
+                .assertContentType("text/plain");
+    }
+
+    @Test
+    void ifTheObjectMapperFailsReturnEmptyAndAddToParsingError() {
         HttpResponse<RequestCapture> request = Unirest.get(MockServer.INVALID_REQUEST)
                 .asObject(RequestCapture.class);
 
@@ -134,7 +165,7 @@ public class AsObjectTest extends BddTest {
     }
 
     @Test
-    public void ifTheObjectMapperFailsReturnEmptyAndAddToParsingErrorObGenericTypes() {
+    void ifTheObjectMapperFailsReturnEmptyAndAddToParsingErrorObGenericTypes() {
         HttpResponse<RequestCapture> request = Unirest.get(MockServer.INVALID_REQUEST)
                 .asObject(new GenericType<RequestCapture>() {});
 
@@ -146,7 +177,7 @@ public class AsObjectTest extends BddTest {
     }
 
     @Test
-    public void unirestExceptionsAreAlsoParseExceptions() {
+    void unirestExceptionsAreAlsoParseExceptions() {
         HttpResponse<RequestCapture> request = Unirest.get(MockServer.INVALID_REQUEST)
                 .asObject(new GenericType<RequestCapture>() {});
 
@@ -158,7 +189,7 @@ public class AsObjectTest extends BddTest {
     }
 
     @Test
-    public void canSetObjectMapperToFailOnUnknown() {
+    void canSetObjectMapperToFailOnUnknown() {
         com.fasterxml.jackson.databind.ObjectMapper jack = new com.fasterxml.jackson.databind.ObjectMapper();
         jack.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, true);